BZ-1166011 - SessionInfo primary key is Integer

+++ kogito-drools/drools-core/src/main/java/org/drools/core/command/SingleSessionCommandService.java
@@ -18,7 +18,7 @@ package org.drools.core.command;
 
 
 public interface SingleSessionCommandService extends CommandService {
-    int getSessionId();
+    Long getSessionId();
     void dispose();
     void destroy();
 }
